+++ src/restart.h
@@ -25,7 +25,7 @@ pid_t r_wait(int *stat_loc);
 pid_t r_waitpid(pid_t pid, int *stat_loc, int options);
 ssize_t r_write(int fd, void *buf, size_t size);
 ssize_t readblock(int fd, void *buf, size_t size);
-int readline(int fd, char *buf, int nbytes);
+ssize_t readline(int fd, char *buf, ssize_t nbytes);
 ssize_t readtimed(int fd, void *buf, size_t nbyte, double seconds);
 int readwrite(int fromfd, int tofd);
 int readwriteblock(int fromfd, int tofd, char *buf, int size);
